Refactor AddPromptsBookmark components split
