+++ src/js/popup/components/PromptsBookmark/PromptsBookmarkList/index.vue
@@ -25,38 +25,21 @@
       @dragenter.prevent
     />
   </div>
-  <div class="add-bookmark-modal-open-button-container">
-    <button
-      v-if="promptsBookmarkArray.length < promptsBookmarksLimited"
-      @click="openAddBookmarkModal"
-    >
-      + Add Bookmark ({{ promptsBookmarkArray.length }} / {{ promptsBookmarksLimited }})
-    </button>
-    <button
-      v-if="promptsBookmarkArray.length >= promptsBookmarksLimited"
-      disabled
-    >
-      Up to {{ promptsBookmarksLimited }} bookmarks...
-    </button>
-  </div>
-  <AddPromptsBookmarkModal 
-    v-if="visibleAddBookmarkModal"
-    :close="closeAddBookmarkModal"
+  <AddPromptsBookmark
     :promptsBookmarkArray="promptsBookmarkArray"
-    :addedCallBack="addedPromptsBookmarkCallback"
   />
 </template>
  
 <script>
 
-import AddPromptsBookmarkModal from "~/js/popup/components/common/AddPromptsBookmarkModal/index.vue";
 import PromptsBookmarkItem from "./PromptsBookmarkItem/index.vue";
+import AddPromptsBookmark from "./AddPromptsBookmark/index.vue";
 import { useEnhancer } from "./enhancer";
  
 export default {
   components: {
-    AddPromptsBookmarkModal,
-    PromptsBookmarkItem
+    PromptsBookmarkItem,
+    AddPromptsBookmark
   },
   props: {
     filterWord: {
@@ -74,28 +57,4 @@ export default {
   .prompts-bookmark-list {
     margin-bottom: 40px;
   }
-
-  .add-bookmark-modal-open-button-container {
-    position: fixed;
-    bottom: 12px;
-    left: 0px;
-    right: 0px;
-    margin: auto 8px;
-
-    button {
-      width: 100%;
-      font-size: 14px;
-      font-weight: 700;
-      padding: 4px;
-      background: white;
-      border: 1px solid gray;
-      text-align: center;
-      cursor: pointer;
-
-      &:disabled {
-        background: silver;
-      }
-    }
-  }
-  
 </style>
